rename the busl package to publisher
Having a command and package named busl can be quite misleading as to
where something should be.
This rename makes it clearer this package is the lower-level interface
allowing publishing data.
